apps/api/pyproject.toml
[project]
name = "airis-mcp-gateway-api"
version = "2.1.0"
description = "AIRIS MCP Gateway API - Fixes Docker MCP Gateway initialized notification"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
    "pydantic>=2.10.0,<3.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "psutil>=5.9.0,<7.0.0",
    "cryptography>=46.0.6,<48.0.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-cov>=4.0.0,<7.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "aiosqlite>=0.20.0,<1.0.0",
    "greenlet>=3.0.0,<4.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]
